Consolidate application code
